requirements.txt

fastapi>=0.111.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-jose
passlib[bcrypt]
uvicorn
python-dotenv==1.1.0
ruff
pytest
google-generativeai
httpx
